improve transform string performance
